Created a new package directory for counter and moved counter.py, FeatureSelector.py, hts_parsing.py, and statistics.py to it since these are all counter-related resources. Cleaned up imports a bit by using package relative imports where possible.
Created util.py for utility functions that may be shared across pipeline steps. The 'from_here' function has been moved there.

In counter, also made changes to prevent duplicate Name Attributes from being added for a GFF file (oops). Made a minor change to the argparse variable. Before it shadowed the global namespace variable 'parser'
