+++ setup.py
@@ -37,7 +37,7 @@ setuptools.setup(
         'console_scripts': [
             'aquatx = aquatx.aquatx:main',
             'aquatx-config = aquatx.srna.Configuration:Configuration.main',
-            'aquatx-collapse = aquatx.srna.collapser:main',
+            'aquatx-collapse = aquatx.srna.collapser.collapser:main',
             'aquatx-count = aquatx.srna.counter:main'
         ]
     },
